refactor: combine aggregation related class into a single file
