+++ stupidb/aggregator.py
@@ -1,9 +1,12 @@
 """Base aggregator interface."""
 
-import abc
-from typing import Generic, Optional, Sequence, TypeVar
+from __future__ import annotations
 
-from .typehints import Result, T
+import abc
+from typing import Callable, ClassVar, Generic, Optional, Sequence, Tuple, TypeVar
+
+from .row import AbstractRow
+from .typehints import Getter, Output, Result, T
 
 AggregateClass = TypeVar("AggregateClass", covariant=True)
 
@@ -13,7 +16,7 @@ class Aggregator(Generic[AggregateClass, Result], abc.ABC):
 
     Aggregators must implement the :meth:`~stupidb.aggregator.Aggregator.query`
     method. Aggregators are tied to a specific kind of aggregation. See the
-    :meth:`~stupidb.aggregatetypes.UnaryAggregate.prepare` method for how to
+    :meth:`~stupidb.aggregator.Aggregate.prepare` method for how to
     provide a custom aggregator.
 
     See Also
@@ -30,3 +33,23 @@ class Aggregator(Generic[AggregateClass, Result], abc.ABC):
     @abc.abstractmethod
     def query(self, begin: int, end: int) -> Optional[Result]:
         """Query the aggregator over the range from `begin` to `end`."""
+
+
+class Aggregate(Generic[Output], abc.ABC):
+    """An aggregate or window function."""
+
+    __slots__ = ()
+    aggregator_class: ClassVar[Callable[..., Aggregator]]
+
+    @classmethod
+    def prepare(
+        cls,
+        possible_peers: Sequence[AbstractRow],
+        getters: Tuple[Getter, ...],
+        order_by_columns: Sequence[str],
+    ) -> Aggregator[Aggregate[Output], Output]:
+        """Prepare an aggregation of this type for computation."""
+        arguments = [
+            tuple(getter(peer) for getter in getters) for peer in possible_peers
+        ]
+        return cls.aggregator_class(arguments, cls)
